fazle-system/api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
httpx==0.28.1
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
PyJWT[crypto]==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
psycopg2-binary==2.9.10
livekit-api==0.7.1
prometheus-fastapi-instrumentator==7.0.2
PyPDF2==3.0.1
python-docx==1.1.2
Pillow==11.1.0
minio==7.2.12
openai==1.58.1
